requirements.txt
torch --index-url https://download.pytorch.org/whl/cpu
torchaudio --index-url https://download.pytorch.org/whl/cpu
tensorflow
demucs
basic-pitch
yt-dlp
spotdl
flask
pydub
python-slugify